Fix propdeps duplicate eclipse classpath
Prior to this commit the 'provided' and 'optional' configurations were
explicitly added to the eclipse classpath 'plusConfigurations'. This
could cause duplicate classpath entries when different versions of
an artifact appeared in more than one configuration.

The eclipse plugin now only adds the 'testRuntime' configuration which
is modified to extend 'provided' and 'optional'.